requirements.txt
streamlit
openai
pinecone-client
google-cloud-storage
pdfplumber
python-docx